+++ app/_template/README.md
@@ -1,0 +1,3 @@
+# Template Directory
+
+This directory contains all template-specific code for the Clerk + Next.js starter template. It's designed to be easily removable when you're ready to build your own application.
