+++ app/_template/components/clerk-logo.tsx
