+++ app/_template/components/footer.tsx
@@ -109,7 +109,7 @@ export function Footer() {
                 </clipPath>
               </defs>
             </svg>
-            X (formally Twitter)
+            X (formerly Twitter)
           </a>
         </li>
         <li>
