--- app/sign-up/[[...sign-up]]/page.tsx
@@ -1,9 +1,0 @@
-import { SignUp } from "@clerk/nextjs";
-
-export default function Page() {
-  return (
-    <div className="flex justify-center py-24">
-      <SignUp />
-    </div>
-  );
-}
